// Frontend/src/app/generated-api/model/wallpaperResponse.ts
/**
 * Device Manager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Represents an available wallpaper
 */
export interface WallpaperResponse { 
    /**
     * Unique identifier of the wallpaper
     */
    id?: string;
    /**
     * Name of the wallpaper
     */
    name?: string;
}