// Frontend/src/app/generated-api/model/themeResponse.ts
/**
 * Device Manager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Represents an available theme
 */
export interface ThemeResponse { 
    /**
     * Unique identifier of the theme
     */
    id?: string;
    /**
     * Name of the theme
     */
    name?: string;
}