// Frontend/src/app/generated-api/model/addApplicationRequest.ts
/**
 * Device Manager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request body for adding an application to a menu or submenu
 */
export interface AddApplicationRequest { 
    /**
     * Identifier of the application to add
     */
    applicationId: string;
}