// Frontend/src/app/generated-api/model/createSubMenuRequest.ts
/**
 * Device Manager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request body for creating a new submenu
 */
export interface CreateSubMenuRequest { 
    /**
     * Name of the submenu
     */
    name: string;
}