// Frontend/src/app/generated-api/model/subMenuResponse.ts
/**
 * Device Manager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Short summary of a submenu
 */
export interface SubMenuResponse { 
    /**
     * Unique identifier of the submenu
     */
    id?: string;
    /**
     * Name of the submenu
     */
    name?: string;
}